sway: delay pme startup to try to avoid potential race with mako
pme sometimes failed to show the notification on timeout, and this issus
seems occurs only after monitor layout has changed which seems to be
fixed by adding `makoctl reload` to kanshi config.
But it is still reasonable to delay pme startup, to avoid too many execs
at the same time.